requirements.txt
pycrypto
enum34